Ahora usamos tenant_id y no empresa

+++ eliminar_usuario.py
@@ -9,18 +9,21 @@ table = dynamodb.Table(table_name)
 def lambda_handler(event, context):
     try:
         body = json.loads(event['body'])
-        tenant_id = body['tenant_id']
-        user_id = body['id']
+        tenant_id = body.get('tenant_id')
+        user_id = body.get('id')
+
+        if not tenant_id or not user_id:
+            return {'statusCode': 400, 'body': json.dumps({'message': 'Faltan tenant_id o id en la solicitud'})}
 
         response = table.delete_item(
             Key={'tenant_id': tenant_id, 'id': user_id},
-            ConditionExpression='attribute_exists(id)'  # Para evitar eliminar un registro inexistente
+            ConditionExpression='attribute_exists(id)'
         )
 
-        return {'statusCode': 200, 'body': json.dumps({'message': 'Usuario eliminado correctamente'})}
+        return {'statusCode': 200, 'body': json.dumps({'message': f'Usuario {user_id} eliminado correctamente'})}
 
     except dynamodb.meta.client.exceptions.ConditionalCheckFailedException:
-        return {'statusCode': 404, 'body': json.dumps({'message': 'Usuario no encontrado'})}
+        return {'statusCode': 404, 'body': json.dumps({'message': 'No se encontró el usuario para eliminar'})}
 
     except Exception as e:
-        return {'statusCode': 400, 'body': json.dumps({'error': str(e)})}
+        return {'statusCode': 500, 'body': json.dumps({'message': 'Error al eliminar usuario', 'error': str(e)})}
